Add base project. StartVC, PlayVC, GameOverVC.
Make the transition from StartVC to PlayVC and the transition from GameOverVC to StartVC and PlayVC. Make a Views for the background, windows and cucumber jars. Animate the background and windows using a timer. Add a pan gesture recognizer for a jar of cucumbers. Add a check for the intersection of a jar with cucumbers and windows. Add scoring.